+++ src/main/java/ie/wellbeing/repository/PaymentDetailsDao.java
@@ -13,6 +13,6 @@ import java.util.List;
 
 @Repository
 public interface PaymentDetailsDao extends JpaRepository<PaymentDetails, Integer> {
-    List<PaymentDetails> findByPaymentUserId(Integer id);
+    List<PaymentDetails> findByPaymentUserId(Long id);
 
 }
